% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anecdotes1.R
\docType{data}
\name{data}
\alias{data}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{id}{}
\item{subject}{}
\item{datetime}{}
\item{alignment}{}
\item{anecdote}{}
\item{contact}{}
\item{sex}{}
\item{age}{}
\item{other}{}
\item{business_exp}{}
\item{business_edu}{}
\item{total_time}{}
\item{allocation_projectA}{}
\item{allocation_projectB}{}
\item{follow_up_similarity_rating}{}
\item{follow_up_similarity_justify}{}
\item{follow_up_relevance_specific_rating}{}
\item{follow_up_relevance_specific_justify}{}
\item{follow_up_relevance_general_rating}{}
\item{follow_up_relevance_general_justify}{}
\item{sample}{}
\item{date_rank}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{anecdotes1}
}
