% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_responses.R
\name{get_responses}
\alias{get_responses}
\title{Extract responses from jsPsych data}
\usage{
get_responses(data, names_to, names_pattern, values_to = "value")
}
\arguments{
\item{values_to}{}
}
\value{

}
\description{
Extract responses from jsPsych data
}
