% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_anecdotes_1.R
\name{import_data_anecdotes_1}
\alias{import_data_anecdotes_1}
\title{Import anecdotes 1

Data was collected through both Pavlovia and the psych server}
\usage{
import_data_anecdotes_1(data_directory)
}
\value{

}
\description{
Import anecdotes 1

Data was collected through both Pavlovia and the psych server
}
\author{
Shir Dekel
}
