% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anecdotes1.R
\docType{data}
\name{power}
\alias{power}
\title{Descriptive data title}
\format{
a \code{list} containing the following fields:
\describe{
\item{anova_n}{}
\item{manova_n}{}
\item{emm_n}{}
\item{plot_ANOVA}{}
\item{plot_MANOVA}{}
\item{plot_emm}{}
\item{power_df}{}
\item{power_df_manova}{}
\item{power_df_emm}{}
\item{effect_sizes}{}
\item{effect_sizes_manova}{}
\item{effect_sizes_emm}{}
\item{n}{}
\item{n_total}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{anecdotes1}
}
