% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anecdotes1.R
\docType{package}
\name{anecdotes1}
\alias{anecdotes1}
\alias{anecdotes1-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
anecdotes1
A data package for anecdotes1.

Use \code{data(package='anecdotes1')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
}
